fixing install.sh and index.html

+++ install.sh
@@ -13,7 +13,12 @@ sudo cp -r website/* /var/www/portfolio/
 sudo chown -R www-data:www-data /var/www/portfolio
 
 # Salin konfigurasi nginx
+if [ ! -f /etc/nginx/sites-available/portfolio ]; then
+echo "⚙️ Copying nginx.conf to /etc/nginx/sites-available/portfolio..."
 sudo cp nginx.conf /etc/nginx/sites-available/portfolio
+else
+echo "⚠️ nginx.conf sudah ada di server, tidak di-overwrite."
+fi
 if [ ! -L /etc/nginx/sites-enabled/portfolio ]; then
     sudo ln -s /etc/nginx/sites-available/portfolio /etc/nginx/sites-enabled/
 else
